Update README.md
This home appliance company serves customers in 3 main product types: refrigerator, washing machine and dishwasher. Company employees need to know quickly what products are in the inventory, what features these products have, and how many of them in the inventory are in order to better serve customers. For this reason, inventory program is very important. The program first starts by listing the products in the inventory. Then Program asks to the employee; if you want to add a product  or make a change in the number of products in the inventory, and program work according to this question.
 All machines in the Inventory Program have the same and different propertiesIn order to prevent the same properties from increasing the file size, the C ++ language must be written using the Inheritance  and Abstraction property. For this reason, a base class named Machine was created; and this class includes an identification code, height, length, and width values in centimeters to represent the dimensions of the product and energy consumption class, which is represented as a set of characters, to specify energy consumption of the product, which are common properties to all machines. At the same time, each machine has a motor, which features power in watts and maximum rotational speed per minute. Thus, all the above mentioned properties can be used for all machine types. Some of these properties are of type 'int' and some of them are of type 'string', so 'int to string' conversion is required. 
After the creation of the base class, subclasses were created according to the different properties of the machines. These subclasses have the same names as the type of machines in the inventory. In these subclasses, the vector container and of the STL library is used.  Because containers classes store objects and data and vector has the ability to resize itself automatically when an element is inserted or deleted. Since the main purpose of the program is to make a change in the products in the inventory, the use of the STL library is important because it allows the changes to be made and performs these operations very quickly. At the same time, the elements of the vectors are quickly reached and the complexity of the program is reduced thanks to the iterators. The 'string tokenizer' function was used to separate the machine features entered in sequence and assign them to the properties to which they belong. The 'fstream' library was used to list the products in the inventory, to make a change in the number of products in the inventory, or to add a new product to the inventory. This library has functions for reading data from text and writing data to text (ifstream, ofstream).
The refrigerator class, the first of the subclasses, have minimum temperature, volume in liters, number of doors and defrost properties. The washing machine subclass have water consumption in liters, number of cleaning programs, capacity in kilograms, and squeeze speed properties. The dishwasher subclass have water consumption in liters, number of cleaning programs, capacity as number of dishes, and number of racks properties. All these properties are assigned by the set functions to strings separated  using the 'tokenizer' structure. 
To add a product to the inventory, the program first asks which type of product you want to add, and then asks for all parameters for the product to be added. The program adds this product to the inventory and exits the program. 
In the process of changing the number of a product in the inventory, the program first asks for the product type to be changed, then lists the IDs of that product type, and then asks for the ID of the product to be changed, then all parameters are entered with a changed inventory number. The program writes it to the inventory and exits the program. NOTE: The number that indicates the number of products in the inventory is the last number of the products listed.